requirements.txt
flask
Pillow==8.3.2
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
nb-black==1.0.7
numpy==1.20.1
matplotlib==3.3.4
seaborn==0.11.1